BottomNavigationView перенсен во фрагмент fragment_notes_list. Получилось добить наконец-то переходы между фрагментами
