+++ app/src/main/java/mynotes/mynotes/MainActivity.java
@@ -8,13 +8,9 @@ import androidx.fragment.app.FragmentTransaction;
 
 import android.os.Bundle;
 import android.view.MenuItem;
-import android.view.View;
-import android.widget.Toast;
 
 import com.google.android.material.bottomnavigation.BottomNavigationView;
-import com.google.android.material.floatingactionbutton.FloatingActionButton;
 
-import mynotes.mynotes.domain.Note;
 import mynotes.mynotes.ui.item_note.ItemNoteFragment;
 import mynotes.mynotes.ui.notes_list.NotesFragmentList;
 
@@ -25,40 +21,11 @@ public class MainActivity extends AppCompatActivity {
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
-
-        BottomNavigationView navView = findViewById(R.id.nav_view);
-
-        navView.setOnNavigationItemSelectedListener(new BottomNavigationView.OnNavigationItemSelectedListener() {
-            @Override
-            public boolean onNavigationItemSelected(@NonNull MenuItem item) {
-                int itemId = item.getItemId();
-
-                if (itemId == R.id.btn_notes) {
-                    openTab(new NotesFragmentList(), NotesFragmentList.TAG);
-                    return true;
-                } else if (itemId == R.id.btn_notes_grid) {
-                    openTab(new ItemNoteFragment(), ItemNoteFragment.TAG);
-                    return true;
-                }
-                return false;
-            }
-        });
-
-
         if (savedInstanceState == null) {
-            openTab(new NotesFragmentList(), NotesFragmentList.TAG);
+            FragmentManager fragmentManager = getSupportFragmentManager();
+            FragmentTransaction fragmentTransaction = fragmentManager.beginTransaction();
+            fragmentTransaction.replace(R.id.fragment_main, new NotesFragmentList(), NotesFragmentList.TAG);
+            fragmentTransaction.commit();
         }
     }
-
-    private void openTab(Fragment fragment, String tag) {
-        Fragment addedFragment = getSupportFragmentManager().findFragmentByTag(tag);
-
-        if (addedFragment == null) {
-            getSupportFragmentManager()
-                    .beginTransaction()
-                    .replace(R.id.fragment_main, fragment, tag)
-                    .commit();
-        }
-    }
-
 }
