+++ app/src/main/java/mynotes/mynotes/ui/item_note/ItemNoteFragment.java
@@ -5,11 +5,13 @@ import android.os.Bundle;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
+import androidx.lifecycle.Observer;
 import androidx.lifecycle.ViewModelProvider;
 
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import android.widget.TextView;
 
 import mynotes.mynotes.R;
 
@@ -20,19 +22,30 @@ public class ItemNoteFragment extends Fragment {
 
     private ItemNoteViewModel itemNoteViewModel;
 
+    @Override
+    public void onCreate(@Nullable Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        itemNoteViewModel = new ViewModelProvider(this).get(ItemNoteViewModel.class);
 
-    public View onCreateView(@NonNull LayoutInflater inflater,
-                             ViewGroup container, Bundle savedInstanceState) {
-        itemNoteViewModel =
-                new ViewModelProvider(this).get(ItemNoteViewModel.class);
-        View root = inflater.inflate(R.layout.fragment_item_note, container, false);
 
-        return root;
     }
+
+    @Nullable
+    @Override
+    public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
+
+        return inflater.inflate(R.layout.fragment_item_note, container, false);
+
+    }
+
+
+
+
 
     @Override
     public void onViewCreated(@NonNull View view, @Nullable Bundle savedInstanceState) {
         super.onViewCreated(view, savedInstanceState);
-
+        TextView textView = view.findViewById(R.id.input_text_note);
+        TextView textView2 = view.findViewById(R.id.input_name_note);
     }
 }
