+++ app/src/main/java/mynotes/mynotes/ui/notes_list/NotesFragmentList.java
@@ -9,14 +9,17 @@ import androidx.fragment.app.FragmentManager;
 import androidx.fragment.app.FragmentTransaction;
 import androidx.lifecycle.Observer;
 import androidx.lifecycle.ViewModelProvider;
+import androidx.recyclerview.widget.GridLayoutManager;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
 import android.view.LayoutInflater;
+import android.view.MenuItem;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Toast;
 
+import com.google.android.material.bottomnavigation.BottomNavigationView;
 import com.google.android.material.floatingactionbutton.FloatingActionButton;
 
 import java.util.List;
@@ -60,7 +63,7 @@ public class NotesFragmentList extends Fragment {
 
         RecyclerView noteList = view.findViewById(R.id.notes_list);
         noteList.setAdapter(adapter);
-        noteList.setLayoutManager(new LinearLayoutManager(requireContext()));
+//        noteList.setLayoutManager(new LinearLayoutManager(requireContext()));
 
 
         notesViewModel.getNotesLiveData()
@@ -77,12 +80,39 @@ public class NotesFragmentList extends Fragment {
         floatingActionButton.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                FragmentManager fragmentManager = requireActivity().getSupportFragmentManager();
-                FragmentTransaction fragmentTransaction = fragmentManager.beginTransaction();
-                fragmentTransaction.replace(R.id.fragment_main, new ItemNoteFragment());
-                fragmentTransaction.addToBackStack(null);
-                fragmentTransaction.commit();
+                openTab(new ItemNoteFragment(), ItemNoteFragment.TAG);
             }
         });
+
+        BottomNavigationView navView = view.findViewById(R.id.nav_view);
+
+        navView.setOnNavigationItemSelectedListener(new BottomNavigationView.OnNavigationItemSelectedListener() {
+            @Override
+            public boolean onNavigationItemSelected(@NonNull MenuItem item) {
+                int itemId = item.getItemId();
+                openTab(new NotesFragmentList(), NotesFragmentList.TAG);
+                if (itemId == R.id.btn_notes) {
+                    noteList.setLayoutManager(new LinearLayoutManager(requireContext()));
+                    return true;
+                } else if (itemId == R.id.btn_notes_grid) {
+                    noteList.setLayoutManager(new GridLayoutManager(requireContext(), 3));
+                    return true;
+                }
+                return false;
+            }
+        });
+
+    }
+
+    private void openTab(Fragment fragment, String tag) {
+        Fragment addedFragment = requireActivity().getSupportFragmentManager().findFragmentByTag(tag);
+
+        if (addedFragment == null) {
+            requireActivity().getSupportFragmentManager()
+                    .beginTransaction()
+                    .replace(R.id.fragment_main, fragment, tag)
+                    .commit();
+        }
     }
 }
+
