+++ app/src/main/res/values/strings.xml
@@ -4,4 +4,6 @@
     <string name="hello_blank_fragment">Hello blank fragment</string>
     <string name="settings">Настройки</string>
     <string name="info">О приложении</string>
+    <string name="title_note">Тема заметки</string>
+    <string name="text_note">Текст заметки</string>
 </resources>
